update: topic name USER_TICKET_QUANTITY_ADD_FAILED -> TICKET_ADDITION_FAILED change
