+++ src/main/kotlin/gogo/gogoshop/global/kafka/consumer/dto/TicketAdditionFailedEvent.kt
@@ -2,7 +2,7 @@ package gogo.gogoshop.global.kafka.consumer.dto
 
 import gogo.gogoshop.domain.shop.receipt.persistence.TicketType
 
-data class UserTicketQuantityAddFailedEvent(
+data class TicketAdditionFailedEvent(
     val id: String,
     val miniGameId: Long,
     val ticketType: TicketType,
