+++ src/main/kotlin/gogo/gogoshop/global/kafka/properties/KafkaTopics.kt
@@ -2,7 +2,6 @@ package gogo.gogoshop.global.kafka.properties
 
 object KafkaTopics {
     const val TICKET_SHOP_BUY = "ticket_shop_buy"
-    const val TICKET_SHOP_BUY_FAILED = "ticket_shop_buy_failed"
     const val TICKET_POINT_MINUS_FAILED = "ticket_point_minus_failed"
-    const val USER_TICKET_QUANTITY_ADD_FAILED = "user_ticket_quantity_add_failed"
+    const val TICKET_ADDITION_FAILED = "ticket_addition_failed"
 }
